requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
